src/drift/client.py
"""
Drift Protocol API接続モジュール
"""
import os
import json
import time
from loguru import logger
from solana.rpc.api import Client
from solana.keypair import Keypair
from anchorpy import Provider, Wallet
from solana.rpc.types import TxOpts
from solana.publickey import PublicKey
import base58

class DriftClient:
    """Drift Protocolとの接続・操作を行うクライアントクラス"""
    
    def __init__(self, config=None):
        """
        Drift Protocolクライアントの初期化
        
        Args:
            config (dict, optional): 設定情報。指定がない場合は環境変数から読み込み
        """
        # 設定の読み込み
        self.config = config or {}
        self.rpc_url = self.config.get('rpc_url') or os.getenv('SOLANA_RPC_URL')
        self.private_key_path = self.config.get('private_key_path') or os.getenv('SOLANA_PRIVATE_KEY_PATH')
        
        # Solana RPCクライアントの初期化
        self.solana_client = Client(self.rpc_url)
        
        # キーペアの読み込み
        self.keypair = self._load_keypair()
        
        # Anchorプロバイダーの設定
        self.wallet = Wallet(self.keypair)
        self.provider = Provider(self.solana_client, self.wallet, opts=TxOpts(skip_preflight=True))
        
        logger.info(f"Drift Protocol client initialized with address: {self.keypair.public_key}")
    
    def _load_keypair(self):
        """
        Solanaキーペアを読み込む
        
        Returns:
            Keypair: Solanaキーペア
        """
        try:
            with open(self.private_key_path, 'r') as f:
                keypair_data = json.load(f)
                return Keypair.from_secret_key(bytes(keypair_data))
        except Exception as e:
            logger.error(f"Failed to load Solana keypair: {e}")
            raise
    
    async def get_funding_rate(self, market="BTC-PERP"):
        """
        指定された市場のファンディングレートを取得
        
        Args:
            market (str): 市場シンボル
            
        Returns:
            float: ファンディングレート（8時間ごとのレート）
        """
        # 実際の実装ではDrift ProtocolのAPIを使用してファンディングレートを取得
        # このサンプルでは仮の実装
        logger.info(f"Getting funding rate for {market} from Drift Protocol")
        
        # TODO: 実際のDrift Protocol APIを使用してファンディングレートを取得する実装に置き換え
        # 現在は仮の値を返す
        return 0.0001  # 仮の値
    
    async def get_position(self, market="BTC-PERP"):
        """
        指定された市場のポジション情報を取得
        
        Args:
            market (str): 市場シンボル
            
        Returns:
            dict: ポジション情報
        """
        # 実際の実装ではDrift ProtocolのAPIを使用してポジション情報を取得
        # このサンプルでは仮の実装
        logger.info(f"Getting position for {market} from Drift Protocol")
        
        # TODO: 実際のDrift Protocol APIを使用してポジション情報を取得する実装に置き換え
        # 現在は仮の値を返す
        return {
            "size": 0.0,
            "entry_price": 0.0,
            "liquidation_price": 0.0,
            "margin": 0.0,
            "unrealized_pnl": 0.0
        }
    
    async def open_position(self, market="BTC-PERP", side="long", size=0.0, price=None):
        """
        指定された市場でポジションを開く
        
        Args:
            market (str): 市場シンボル
            side (str): 取引方向 ("long" または "short")
            size (float): ポジションサイズ（USD）
            price (float, optional): 指値価格。Noneの場合は成行注文
            
        Returns:
            dict: 注文結果
        """
        # 実際の実装ではDrift ProtocolのAPIを使用してポジションを開く
        # このサンプルでは仮の実装
        logger.info(f"Opening {side} position for {size} USD in {market} on Drift Protocol")
        
        # TODO: 実際のDrift Protocol APIを使用してポジションを開く実装に置き換え
        # 現在は仮の値を返す
        return {
            "order_id": "sample_order_id",
            "status": "filled",
            "filled_size": size,
            "average_price": 50000.0  # 仮の値
        }
    
    async def close_position(self, market="BTC-PERP", side="long"):
        """
        指定された市場のポジションを閉じる
        
        Args:
            market (str): 市場シンボル
            side (str): 取引方向 ("long" または "short")
            
        Returns:
            dict: 注文結果
        """
        # 実際の実装ではDrift ProtocolのAPIを使用してポジションを閉じる
        # このサンプルでは仮の実装
        logger.info(f"Closing {side} position in {market} on Drift Protocol")
        
        # TODO: 実際のDrift Protocol APIを使用してポジションを閉じる実装に置き換え
        # 現在は仮の値を返す
        return {
            "order_id": "sample_close_order_id",
            "status": "filled",
            "filled_size": 0.0,
            "average_price": 50000.0  # 仮の値
        }
    
    async def get_account_balance(self):
        """
        アカウント残高を取得
        
        Returns:
            float: アカウント残高（USD）
        """
        # 実際の実装ではDrift ProtocolのAPIを使用してアカウント残高を取得
        # このサンプルでは仮の実装
        logger.info("Getting account balance from Drift Protocol")
        
        # TODO: 実際のDrift Protocol APIを使用してアカウント残高を取得する実装に置き換え
        # 現在は仮の値を返す
        return 1000.0  # 仮の値
